Update patch set 1
Patch Set 1:

> It does prevent the device from booting again next time, which is also accomplished by verifying the entire partition once at boot.

In this case it would probably be easier to just use an AVB hash-descriptor instead of a hashtree-descriptor (and have fs_mgr support this). Would save space (no hashtree needed) and probably a few cycles of computation time! But as Will says, maybe dm-verity all-the-time is fast/efficient enough for the device in question!

Patch-set: 1