Update patch set 8
Patch Set 8:

(1 comment)

Testing this change of heart regarding android_set_log_frontend.

NB: LOGGER_FIFO is another mask that was in part abandoned because we could not achieve the performance we were looking for. That one spawned a background thread and created a produce-consumer near lockless queue to feed it so as to reduce logging overhead on the source of the information. We will always hope we can find an algorithm that makes it worth it ...

Patch-set: 8