init: don't treat android.selinux=disabled as enforcing
When cmdline key android.selinux is set as disabled, current code
will treat that as enforcing and that mode will be set to kernel.

I'm not sure if "disable" is still a valid mode, but it is still in the
offical doc[1] at this moment. At the least, this mode is useful in
development stage.

[1]https://source.android.com/security/selinux/validate.html

Change-Id: Icab064c2e404f8028fc88c5931b9cb0d48939deb
Signed-off-by: Bin Chen <bin.chen@linaro.org>