Update patch set 1
Patch Set 1:

HI, Nick. Thanks for the review.
 
> Android won't boot if SELinux isn't enabled. 
I'm currently running Android N on 3.18 kernel with SELinux disabled. It can boot to init, and due to the logic as explained in the comment, init will set SELinux as enforcing. And since the seplicy hasn't been setup properly yet, there are many svc denials to the core system services. But it can be fixed by `su setenforce 0`. 

The system still boot, to a certain extend but won't boot to UI. Is this behavior by design? 

 > The only acceptable
 > modes for this are permissive or enforcing. The documentation
 > should be fixed instead.
OK. Let me see if I can fire a patch.

Patch-set: 1