Update patch set 4
Patch Set 4:

(2 comments)

> (2 comments)
 > 
 > i'm not sure this is really adding anything. isn't this just
 > 
 > writepid ${ro.cpuset.default}
 > 
 > ?

No, not really. If cpuset is not specified, then ro.cpuset.default is used. And that's the primary motivation for this additional change. If .rc script uses writepid instead of cpuset, then the process might be put into the default system cpuset first and then maybe hopefully moved to the specified cpuset (based on the order of how the options are applied once the process gets forked).

This latest change will detect that writepid is actually writing to cpuset and as such will not try to use ro.cpuset.default at all...

Patch-set: 4