Fix-up coding style and add PREUPLOAD.cfg file.
The .clang-format is a symlink to ../init/.clang-format, which is merged
recently. As init is the major user of fs_mgr, it's better to keep the
style consistent.

Finally, add a PREULOAD.cfg file to do this on every commit. It only runs
clang-format on the diff lines the uploading commit. See the following
links for more information:

  https://android.googlesource.com/platform/tools/repohooks/

Bug: None
Test: repo upload, checks it only formats the diff lines
Test: Device can boot with AVB

Change-Id: I5c72f23b38534d5bcef3e4f9f0f477fa40496433