Update patch set 3
Patch Set 3:

Serialization is just a simple memory copy (I am not addressing any issue with endianess).
Also I needed to create a copy of the data buffer inside fb_queue_partition that is not free right now. Since std::vector<uint8_t> data inside do_partition_table seems to get out of scope before the buffer is sent and client side it getting garbage. I tried to move it data as a variable of main function so it will be destroyed when main ends but it did not help. I need to check how to solve that.

Patch-set: 3
Reviewer: Gerrit User 1138338 <1138338@85c56323-6fa9-3386-8a01-6480fb634889>