Update patch set 3
Patch Set 3: Code-Review+1

I'm just wondering if you know if it actually does return a handle and not NULL and/or if that handle is usable since this is before selinux_android_load_policy() and others are called.

The only place sehandle is used, checks for null first, so things will work fine in any case.

I agree there's more refactoring that's needed.  It's scary that this single `sehandle` global is used by both code paths from ueventd and init.

In the end, lgtm; this needs to be refactored wholesale in the future.

Patch-set: 3
Label: Code-Review=+1