liblogcat: Add android::ReadLogcatToString helper
bool android::ReadLogcatToString(const std::string& command,
                                 std::string* content)

Collects output of threaded in-process logcat into content, according
to supplied command.  The input command and output content can
reference the same object to conserve objects.  If the stderr is
redirected to stdout with a "2>&1", will add a trailing "\n#|" to the
end of the content, ala the Android shell behavior, to report the
command exit return value.  Return value is true if there are no
errors in collecting the logcat content.

Test: gTest logcat-unit-tests --gtest_filter=liblogcat.abi_base
Bug: 35326290
Bug: 28035297
Change-Id: I32ac12beccbd12d729dea49913ff91d956c39ee9