Emulator: Enhance qemu_pipe.h to handle partial read and write
Partial read and write happens and it is better to try again
unless there is some hard error. This is meant to fix some
flaky behavior of emulator pipes, hopefully.

TODO: fix this
Copy/Paste the WriteFully and ReadFully from android-base/file.h,
as this qemu_pipe.h is included both by c and cpp file.

BUG: 35207286

manually tested this on emulator image.

Change-Id: If033c815e621002313c45d1bcd9a92d2d75da4ad