Update patch set 1
Patch Set 1:

In general, we are setting the gcov configurations at test run-time. This change is intended to fix the case when gcov is unintentionally invoked. Another team is using the same coverage-enabled target (but not for gcov) and the package manager is triggering gcov data to flush. If pm is invoked at the root, then gcov tries to write the files to a read-only directory.

I suppose we could use separate build targets but it seems like that just avoids the problem instead of preventing the bug from happening.

Patch-set: 1
Reviewer: Gerrit User 1121516 <1121516@85c56323-6fa9-3386-8a01-6480fb634889>