Qemu: make the qemu_pipe_open compatible with old apis
Commit c7b098ceb528afc62b1545377201e45f5d37f974 has changed
the qemu_pipe_open interface to require the "pipe:" prefix in
the service name.
However in APIs 24 and before, the "pipe:" prefix is not required
This causes quite some confusion and bugs since it is very common
to forget the difference when working across differnet APIs.

This CL is meant to make qemu_pipe_open work in both cases:
1. if it does begin with 'pipe:', use as is;
2. if the service name does not begin with 'pipe:', add it.

Change-Id: If9782396c03780fad1aadeb8374eb308517dc963