init: don't panic on early mount failure
There is a possible case of partition corruption where the device is
booted into recovery but can never boot because it gets stuck in the
bootloop caused by the mount failure. This breaks the entire usecase for
recovery. Make sure we don't fail if early mount failed. This also
allows legacy devices to continue to be built as they were.

Things that MUST have early mounted partiton e.g. selinux will have
to check for them explicitly instead.

b/35831190

Test: Boot test angler, sideload OTA angler.
Test: Boot test sailfish

Change-Id: I72a85b39355248b179313a5e2fa8911fe83361b3
Signed-off-by: Sandeep Patil <sspatil@google.com>