Update patch set 1
Patch Set 1:

I too checked, although the Condition::wait* callers, as it's really there that needs to be guarded.  Nearly all of them were already wrapped in while loops or had other checks.

https://android-review.googlesource.com/#/c/346166/ is the one area that looks like it needs a check.

It's unfortunate that we can't check the vendor code, but we're not changing the semantics here, just updating the comment.  I hope that if vendor code were liable to having an issue such as this, that it would be fixed if it did reproduce on device.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>