Update patch set 1
Patch Set 1:

I am becoming an apologist for security cargo cult, not something I want to be. This is dealing with an errno that happens to be zero or negative. These system calls should _never_ fail us in this regard for open() or ioctl(), so there is IMHO _no_ cases that this code can trigger. At issue is that if it is zero or negative (imagine a bug where another thread sets value of errno in a race race), the return from the function is erroneous, and could lead to referencing an unexpected file descriptor. Paranoid hardening IMHO. This kind of cargo saves us from little, there is still a bad actor hitting errno, but grants us some better behavior in the face of such bad actors.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>