Update patch set 1
Patch Set 1:

(3 comments)

One thought, tangent from my comments: Is there going to be a case where there're actually 3 sepolicies to be compiled together, one each for system, vendor, and odm?

It seems likely to me that in the full extent of treble-ization that the odm partition will also have its own sepolicy additions and we'll want to handle that as well.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>