Add a 2 width option of clang format.
Make the default the 4 tab space version.

Add a link to the 2 space version for libbacktrace.

The only other difference is that the 2 space tab version allows short
functions on a single line. This is for things like constructors,
short destructors, or accessor functions. I believe this is closer to
the Google C++ style guide.

Make libbacktrace use the 2 width option.

Bug: 36046320

Test: NA
Change-Id: Ida08db18902eeb101f42869dd5590182529d54ef