logd: wakeup wrap timeout if realtime changes drastically
Timeout is the logd internal reference to the --wrap flag in logcat
(issued in the protocol as timeout=7200).  For a nonblocking read with
a selected timeout, the logger waits until either the log buffer is
about to 'wrap' and prune the log entry, or at the specified timeout.
Non blocking in the logger context means that when there are no more
log entries, the socket is closed.  If there is a large user initiated
time change both the log entries and the timeout really get called
into question, so we trigger a release of the logs for clarity.  This
is so that the caller of logcat can deal with what we have.  Caller
may wish to look at the updated local time to figure out any
mitigations, if required.

The logger has a 5 second window for entries to land in time sorted
order into the logging list.  This should offer the caller some
differentiation between precious logging order sequence for
monotonically increasing time, and sequence order in the face of user
initiated time adjustments that break monotonicity.  This CL is about
major (2 hours or more for --wrap) adjustments in time that can cause
FUD in the logging timestream requiring a strong signal rather than
having to do a postmortem.

Test: gTest liblog-unit-tests logcat-unit-test logd-unit-tests
Bug: 35373582
Change-Id: I92cac83be99d68634ffd4ebd2f3a3067cfd0e942