Update patch set 1
Patch Set 1: Code-Review+1

This is essentially a revert of https://android-review.googlesource.com/206455 . See also b/27520850.

This change feels right to me. Most of the devices supported within AOSP already mount debugfs, so this avoids mounting it twice (which causes SELinux denials and is quite frankly, wrong). Also, as jeffv pointed out, this reduces difference between user builds and userdebug builds. As many part of Android rely on debugfs, reducing this skew feels valuable to me.

If we are going to disable debugfs, we should not mount it on all devices, and provide a mechanism for the shell user to mount it conditionally on debuggable devices. paullawrence was working on doing that.

Change LGTM, but adding others to make sure we're not regressing.

Patch-set: 1
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1