Update patch set 2
Patch Set 2:

timeout is the logd internal reference to the --wrap flag in logcat (issued in the protocol as timeout=7200). For a nonblocking read with a selected timeout, the logger waits until either the log buffer is about toe 'wrap' and prune the log entry, or the timeout. If there is a large user initiated time change both the log entries and the timeout really get called into question, so we release the logs. This is so that the caller of logcat can deal with what we have now, and can probe the logs and the local time to figure out a mitigation for the issues surrounding the time change.

Patch-set: 2
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0