Update patch set 3
Patch Set 3: Verified+1

confirmed that this does not _break_ klogd.

Outside the scope of this CL: It should be noted that the marlin runs on rtc time until after the /data is mounted and does not really hit home until the first healthd heartbeat (where logd is triggered to pick up the time environment). On my marlin device rtc time is 01-15-1970. We need to run through the older logs and fix that.

Patch-set: 3
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1