Update patch set 5
Patch Set 5: Code-Review+1

I just went to verify this and I noticed is that read_file() is more restrictive than the current code, as it disallows reads of group or world readable files.

This is likely behavior that we do want, and a check through our code base doesn't look like it will break anything.

But assuming that we do use this behavior, can we add a comment to this CL and README.md indicating this change?

Patch-set: 5
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1