Update patch set 2
Patch Set 2:

Actually, I'm pretty sure I'm not going to submit this. There are very few callers of android_fork_execvp_ext and all of them are fsck related (ext4 or f2fs). It's unlikely that the savings from that single fork + exec are likely to be significant, compared to the amount of time fsck needs.

I'll split out the unit test into a separate change so that we can submit that, and also record some of the numbers in the bug for posterity.

Patch-set: 2
Reviewer: Gerrit User 1014443 <1014443@85c56323-6fa9-3386-8a01-6480fb634889>