Update patch set 1
Patch Set 1:

To clarify: this is not an existing bug, because the lock file doesn't exist yet (it is being added by one of the other changes in the topic).

I looked around for ways of doing this via selinux, but this file is also locked by netmgrd, and there didn't seem to be an easy way to encode that in selinux policies except by changing all the "netmgrd.te" files for all devices. Is there a better way than that?

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>