Update patch set 2
Patch Set 2:

SGTM. I'd prefer if we have some transition time where this code supports both "_a" and "a" (just check if the first char is _?). This value is passed from the bootloader, so to update the command line we need a round-trip to our partners to get an updated bootloader using "a" instead of "_a" and run tests. All Android Things devices (release dev-previews based in N) have bootloaders passing _a so we would need to update all of them.

Patch-set: 2
Reviewer: Gerrit User 1077068 <1077068@85c56323-6fa9-3386-8a01-6480fb634889>