fs_mgr_avb: allow disabling verity when the device is unlocked
To start Treble VTS test, a single AOSP system.img will be flashed onto
the device. The size of AOSP system.img might be different than the
system partition size on device. This makes locating AVB footer fail,
which is assumed to be at the last 64-byte of the partition.

This change allows disabling dm-verity when the device is unlocked and
the HASHTREE_DISABLED flag is set on the top level /vbmeta.

Bug: 35464251
Test: Manually invoke fs_mgr_load_vbmeta_images() and checks
      FS_MGR_SETUP_AVB_HASHTREE_DISABLED when the conditions are met

Change-Id: I3d885719170461208a5c6638101b6c067d8ae7f2