Update patch set 1
Patch Set 1:

The changes to start hwservicemanager earlier got reverted due to some issues that happened overnight, so this is going to have to wait until they're re-added.

I mentioned offline, but I talked with Wei, and we're going to have the issues fixed with qseecomd, so this class can start right after we start hwservicemanager in late-fs, once the above CLs go back in.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>