Update patch set 3
Patch Set 3: Presubmit-Ready+1

My marlin runs after these changes, the build failure appears to be related to other automerger changes, but could be a bullhead failure because of these changes (LogKlog.cpp is sensitive to differences in kernels, but should not result in a bootup failure). Will not ignore, and retrigger TreeHugger to be sure.

Patch-set: 3
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1