Make init builtin command installkey respect property updates
init used to block on intallkey such that it was unable to
process property events. This lead to a deadlock by which
the Keymaster HAL would wait indefinitely for the
hwservicemanager.ready=true. This fixes the issue by
implementing the builtin in terms of do_exec, which allows
init to stay responsive to properties while waiting for the
child to terminate.

Bug: 36278706
Test: Add a 3s delay into hwservicemanager before it sets the property
      hwservicemanager.ready and the device still boots.

Change-Id: Iaefd31156ca01f3a44b4f85a8bf78beee8dfe224