fs_mgr_avb: refactors how vbmeta is loaded
Adds fs_mgr_avb_handle and renames the following functions to work with
the handle:

    - fs_mgr_load_vbmeta_images -> fs_mgr_avb_open
    - fs_mgr_unload_vbmeta_images -> fs_mgr_avb_close

And only invokes fs_mgr_avb_open() when there is a fstab entry having
'avb' flag (need HASHTREE descriptor). fs_mgr_is_avb_used() can be
removed as it only checks system property "ro.boot.vbmeta.hash_alg" to
decide whether vbmeta needs to be loaded, which might not be accurate.

For example, there are only HASH descriptors in the verified chain but
no HASHTREE descriptors. In this case, the fs_mgr doesn't have to do
anything because it only takes care of HASHTREE descriptors.

Bug: 33254008
Test: test AVB on bullhead
Change-Id: Idcb1b0c7f25d9c474fa85ffce2e9a5f38ce7b492