Update patch set 5
Patch Set 5:

> Patch Set 2:
> 
> (1 comment)

NativeBridge version is sort like NDK version of Android API level. I believe, in the development scope of one major version of Android, NDK or API version remains the same one no matter how significant some interfaces or implementations are changed. The reason is simple, these things are not "released" and are unstable.

With this, i think it's reasonable for NativeBridge upgrading to v3 in O rather than v4. And people won't question like "this module is v2 in N and v4 in O, where is v3?"

Nevertheless, if NDK or Android API are so elegant that they don't change ever since in code, NativeBridge should align with other components and upgrade to v4.

Patch-set: 5
Reviewer: Gerrit User 1104175 <1104175@85c56323-6fa9-3386-8a01-6480fb634889>