Update patch set 1
Patch Set 1:

a) assigning to salyzyn@android.com is asking for trouble. Please used salyzyn@google.com
b) jhawkins@google.com is the defacto maintainer, added him.
c) beware of 32-bit tests running on 64 bit platforms, libraries like this will be in /system/lib64/libmetricslogger.so because that is all that is to be delivered, so a 32-bit test will not find /system/lib/libmetricslogger.so

We have the same problem on go/apct for the logcat-unit-test, as liblogcat.so is only supplied as a 64-bit library. See b/34454772 for some details of the struggles.

Patch-set: 1
CC: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>