Update patch set 2
Patch Set 2:

I was primarily looking at the delta from nonencrypted to starting_zygote_secondary, since that's started in class main.

I just re-run the numbers for posterity at TOT on sailfish (20x each):

With this change, I saw
nonencrypted                  : 5.4356  0.13039
starting_zygote_secondary     : 5.5291  0.12954
= 93ms

Without this change, I saw:
nonencrypted                  : 5.407   0.16505
starting_zygote_secondary     : 5.6542  0.12015
= 247ms

So, yes as we figured with this, the delta that we're reducing is between nonencrypted and starting the main class.  This matches the 150ms that you saw and is a bit better than my previous runs.

Full disclosure, With these tests, I did now see boot complete taking slightly longer with the change:

With change:
BootComplete_kernel           : 13.14   0.27874

Without change:
BootComplete_kernel           : 13.043  0.37217

I'm not yet sure the cause or if this is noise, especially with the higher std dev?.

Patch-set: 2