logcatd: introduce logcatd executable
logcatd is the same as logcat, except that the -L flag, if supplied,
runs once, then the command re-runs itself without the -L flag with
the same argument set.  By introducing a logcatd daemon executable
we can solve the problem of the longish reads from pstore that
sometimes occur when the system is excessively busy spinning in a
foreground task starving this daemon as we absorb the delay in
an init service, rather than in an init exec.  This would not have
been efficiently possible without the introduction of liblogcat.

There are no doubt many flags that make no sense to run twice with
and without the -L flag.  In the general sense we expect the caller
to perform the correct set of operations and not explore these less
sensical operations.  logcatd is only supplied on engineering and
debug builds for logpersist as an aid to triage.

Test: gTest logcat-unit-tests
Test: manual confirm logpersist functions as expected
Bug: 28788401
Bug: 30041146
Bug: 30612424
Bug: 35326290
Change-Id: I53ba31970749daf37eef42636f039f485932416f