Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> > > Patch Set 2:
>  > >
>  > > (1 comment)
>  > 
>  > NativeBridge version is sort like NDK version of Android API level.
>  > I believe, in the development scope of one major version of
>  > Android, NDK or API version remains the same one no matter how
>  > significant some interfaces or implementations are changed. The
>  > reason is simple, these things are not "released" and are unstable.
>  > 
>  > With this, i think it's reasonable for NativeBridge upgrading to v3
>  > in O rather than v4. And people won't question like "this module is
>  > v2 in N and v4 in O, where is v3?"
>  > 
>  > Nevertheless, if NDK or Android API are so elegant that they don't
>  > change ever since in code, NativeBridge should align with other
>  > components and upgrade to v4.
> 
> My goal is to not break continuous testing of the platform. I don't think it matters that much what version this ends up being in the final release. Dimitry, wdyt?

Your consideration is absolutely reasonable. That's exactly the reason why we suggested not merge this patch until Houdini has a support for it months ago, and for continuous testing, we even have a plan of enabling "linked namespace" at NativeBridge and Houdini the same time which is being executed.

Since Intel is the only one who utilizes NativeBridge for now, I think it's safe to let it v3 in O while this patch adds a new interface.

Please help to merge this patch together with https://partner-android-review.googlesource.com/c/780478/ which is Houdini's enabling, if you don't mind that NativeBridges is v3.

We can make it v4 if you insist. But it will takes another 2-3 weeks to test NativeBridge and Houdini before we deliver patches, and another 1 weeks to merge. As we still have many other topics to handle, i suggest not this option.

And, thank you for your consideration.

Patch-set: 5
Reviewer: Gerrit User 1104175 <1104175@85c56323-6fa9-3386-8a01-6480fb634889>