Move early_hal start to late-fs trigger action
The class early_hal is essentially for the keymaster hal which needs
to be up before vold tries to unlock a storage encryption key (FDE or
FBE). The current position is too early in the boot process, because
on devices with legacy HAL the wrapper service uses system properties
to find the legacy HAL.

This patch moves the start of the early_hal class to the late-fs trigger
action which runs right after the system property action.

Test: Manually tested and update tested on bullhead and sailfish.
Bug: 35764921

Merged-In: I34b45b85f8450e9ef18861535fdb2ee963df8c9b
Change-Id: I34b45b85f8450e9ef18861535fdb2ee963df8c9b