ziparchive: vectorize ComputeHash
ComputeHash is the hottest function in the profile of libart. This patch
vectorizes the Bernstein's hash function to make it faster: instead of loading
the string byte by byte, we now load 8 bytes at a time and perform the hashing
on all the bytes loaded in the register. The number of loads and branches are
reduced by the vectorization factor SZ. Note that we could use __uint128_t to
increase the vector factor up to 16. This could be done based on further
performance profiling data.

Change-Id: I6b36e74790fe23833d420d796a88cf2d167bfc49