Update patch set 2
Patch Set 2:

> we should make C++ bindings for libavb and only interact with libavb through those bindings.

Just my two cents, but I think this is overkill.

> We can see if it would make sense to include them as a header for libavb itself,

I would be strongly opposed to that. Really, libavb is supposed to be a small self-contained piece of code that's intended to be used in bootloaders. There's also the point that we actually don't promise API stability (we do have promises for on-disk formats, versioning etc.) so we're able to change things when needed.

Patch-set: 2
Reviewer: Gerrit User 1076971 <1076971@85c56323-6fa9-3386-8a01-6480fb634889>