Update patch set 3
Patch Set 3:

Narayan, I think having a benchmark is a very good idea. Thanks for clarifying the comments from Elliott, and thanks for the pointers.

Last Friday I ran the gtests for ziparchive, and these are not representative of the use of ziparchive by libart.

I will first set up the benchmark on x86_64-linux, and if needed after that I can adapt it to run on aarch64-linux Juno and on an Android Exynos-m1 device.

Patch-set: 3
Reviewer: Gerrit User 1155853 <1155853@85c56323-6fa9-3386-8a01-6480fb634889>