init: introduce service group option "shutdown"
The init code currently maintains a list of hard-coded services that
should either:
1. not be killed during reboot at all.
2. not be killed until after apps have been killed.
3. vold, which is started like a shutdown never service,
   but is killed before after_apps is killed.

This patch does not attempt to address #3. Perhaps vold
should handle ShutdownVold() and then it can be moved to
its own thing, like shutdown after_apps_with_start

This poses a few issues:

1. This feature is hard coded into init.
2. Vendor services cannot leverage this.

For item 2, a specific instance in maintaining state between
the IoC and the SoC via CAN bus messages requires that the
service stays alive to send the reboot message from the SoC
to the IoC. where the SoC is the Android device.

Test: reboot the device and check pstore/console-ramoops
      to verify that reboot attempted to start shutdown never
      services like vold as well as stopping after_apps
      services. Ewe can test this by looking for messages like:
         "Starting shutdown never service: <svc name>"
         "Stoping shutdown after_apps service: <svc name>"
      Tested on Hikey, thus watchdogd is not available.

Change-Id: Idac46aec4f14f0fee7675c1a93e4f71869e892d1
Signed-off-by: William Roberts <william.c.roberts@intel.com>